gcc-testsuite: Codex batch — compound literals, comma decls, preprocessor
Codex session work (blocked mid-fix, committing as-is):
- Compound literal detection in more expression contexts
- Comma-separated function declarations: void a(), b();
- Preprocessor: # define with whitespace after #
- sizeof for contextual identifiers
- Fixed-array arrow access
- Various parse improvements

GCC testsuite: 1087 → 1148 passed (+61).

+++ include/datatokens.h
@@ -207,8 +207,8 @@ public:
     // enum/const-var leaves report 0, so e.g. (TOPCOLOR-COLORBASE)
     // folds to 0 at runtime even though parse-time array sizing reads
     // them correctly via read_constant_integer.
-    virtual int ival() const override
-        { return var.is_constant() ? var.get<int>() : 0; }
+    virtual int64_t ival() const override
+        { return var.is_constant() ? var.get<int64_t>() : 0; }
     virtual double dval() const override
         { return var.is_constant() ? var.get<double>() : 0; }
     virtual bool is_constant() const { return var.is_constant(); }
